requirements.txt
streamlit>=1.32.0
google-generativeai>=0.6.0
python-dotenv>=1.0.1
